legal-compliance-agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
faiss-cpu
numpy
pydantic
pdfplumber
python-docx
PyYAML
